Added socket test and reorganized test folder

+++ util/match.js
@@ -44,13 +44,15 @@ const setUserCodeFile = (matchKey, user, code) => {
 
 const deleteFile = (matchKey, user) => {
   let path = `./sessions/${matchKey}_${user}.js`;
-  fs.unlink(path, err => {
-    if (err) {
-      console.log(err);
-      return;
-    }
-    console.log(`${matchKey}_${user}.js file removed`);
-  });
+  if (fs.existsSync(path)) {
+    fs.unlink(path, err => {
+      if (err) {
+        console.log(err);
+        return;
+      }
+      console.log(`${matchKey}_${user}.js file removed`);
+    });
+  }
 };
 
 const putTogetherCodeOnRun = (code, codeConst, expected, test) => {
